refactor: split the ingest module into a package
Everything section 2.7 of the results index plan asks of one ingest
pass carried spindoctor/cli/stats/ingest.py to 1069 lines, past the
1000-line module cap.  It becomes spindoctor/cli/stats/ingest/, split
along the stages a pass runs through:

  counts  the tally the closing summary is read from
  walk    the single listing of a root that feeds every later stage
  store   what the index already holds, and how rows go back in
  chunks  batched retrieval, reading a document, the per-chunk write
  runs    the record that makes absence of a row readable
  driver  the pass itself: walk, select, ingest, prune, complete

Every definition and every module constant moves byte for byte.  No
signature, no name and no behavior changes, and the package re-exports
the surface its consumers already imported, so every existing import
of spindoctor.cli.stats.ingest keeps working unchanged.

The tests that reach into the module follow each patched name to the
module that reads it: a patched retrieval batch size is otherwise
invisible to _ingest_chunk, and a patched _write_image to _write_chunk.
The source scan that enforces criterion 10 finds the split modules for
itself, and the floor it asserts its own reach against names them.
